Wrap FedimintFfi.rpc in setTimeout
It seems like sometimes on Android calls to the bridge block the UI
thread in React Native. Wrapping them in a `setTimeout` will just
await for no time so that Javascript runtime can pass execution
to other functions until bridge RPC call finishes.

This is a hack, but at least it's confined to one place. In future
we should figure out how to make FFI on Android not block.