Refactor bridge listeners
Previously, `addListener` would remove all other listeners for that
`eventType`. This led to a bug where `BitcoinRequest.tsx` would
subscribe to `transaction` events in order to navigate to a the
success page. But in doing so, it unsubscribed the `transaction`
listener on `App.tsx` which would show push notifications.

Now, we just return the use the `EmitterSubscription` returned by
`addListener` to return `() => listener.remove()` in any `useEffect`
which creates a new listener. This way, we only unsubscribe the
specific listener that we subscribed and not any other listeners
for this same event type in other parts of the app.

Also, renamed `FedimintEventEmitter` to `NativeBridgeEventEmitter`
and `TFedimintEventEmitter` to `BridgeEventEmitter`, since the bridge
is now starting to do things unrelated to Fedimint.