bridge: validateRecoveryFile uploads to federation
This fixes a problem where people would have a valid recovery file,
but perhaps for a different federation or something.

Or they would go to the next screen, the upload would fail, and
there would be no way to retry it.

This should probably be renamed `validateAndUploadRecoveryFile`
but that's for a different time!