feat: run android RPC commands in kotlin coroutine
Before this commit, I was noticing that RPC commands on Android could
block UI interactions. By running commands in a Kotlin coroutine that no
longer happens. But we will need to think more about how to actually
cancel the operation in the rust world if the user does navigate away
from the screen.