fix: remove timeout in subscribe_to_ln_pay
this timeout should only exist in the rpc, not subscribe in background.

fixes https://github.com/fedibtc/fedi/issues/3798

There is also bug for overcharging fees because in case of timeout we don't
cancel the payment and payment might succeed later and we add back the
outstanding fee.

This might make fees remit service hang but that is still better than
overcharing. More fixes in follow-ups.