feat: add recovery indicator to ConfirmReceiveOffline
Contributes to https://github.com/fedibtc/fedi/issues/1820

I noticed that if I leave a federation and then join it again from the join foreign
ecash flow, there's a good chance it'll be recovering for a short
period of time. When I tried to claim ecash in the recovery state, I
got an error.

- Redesigns the ConfirmReceiveOffline page
- Removes the "add note" input as you can change it from txn history
  later
- Added a HoloAlert component (from figma designs) to show different
  alerts
- Adds a "you're offline" HoloAlert to the ConfirmReceiveOffline page
  only when you're offline
- Adds a recovery indicator to the ConfirmReceiveOffline page