fix: display svg fedi mod icons correctly (#6195)
Addresses https://github.com/fedibtc/fedi/issues/6174

The Image component in React Native does not support SVGs by default.
The SvgURI component from react-native-svg is required for this.

If a URI ends with SVG, attempts to display it, falling back to the
default fedi mod icon for both Fedi Mods and the preview icon on the
AddFediMod screen.

To test:
- add a URL in the AddFediMod screen
- add the link to an svg image in the icon field
  (e.g. https://tehccringe.com/logo/svg/filled.svg)
- it should display properly
- normal .ico or .png images and other formats should still work
- add a fedi mod that uses an svg as its icon (e.g. tehccringe.com)
- set the icon on the AddFediMod page to be blank
- on the mods page, you should see that the svg icon is auto-fetched