chore: disable certain transaction types if offline
Addresses #5876

If you are offline, shows a banner indicating that you are offline on
the following screens:
- ConfirmSendLightning
- ReceiveLightning
- SendOnChainAmount

Before sending a lighnting payment, creating a lightning invoice,
paying an onchain address, or redeeming an LNURL withdrawal, checks to
ensure that you are indeed online.

If the internet connection is suddenly lost before performing any of
these actions, alerts the user early on that they are offline via a
toast, and then shows the banner.