fix: decryption of multispend events
Alice and Bob are members of a group

sending of first message in matrix encryption session:

1. Alice sends encrypted message in room
2. Alice sends room key (to decrypt that message) using to_device
message to all devices of users in that room (relayed through server).

1 and 2 are happening concurrently.

now imagine the sequence:
- 1 completes
- Bob asks server for new message
- server responds with encrypted message
- Bob is unable to decrypt this encrypted message and saves it as is in
event cache
  - because 2 hasn't completed
- 2 completes
- Bob doesn't update event cache (because event cache doesn't handle
this)

event cache will always return this event as encrypted event

matrix-sdk-ui has retry logic to decrypt the event to handle this
case. Timeline will swap the encrypted message with decrypt message in
background, and you will still see encrypted message for a while.

We also need similar logic in our rescanner because we use event cache
directly. even if we uses the timeline layer, we will need to wait for
decrypt tasks.

references:

https://github.com/matrix-org/matrix-rust-sdk/issues/3872
https://github.com/matrix-org/matrix-rust-sdk/pull/4715
https://github.com/matrix-org/matrix-rust-sdk/issues/4429