chore: improve logic for creating neverthrow errors
The pattern for early-returning a tagged neverthrow error is not verbose
and *very* ugly

```typescript
if(!condition) {
  return err(
    makeError(
      new Error("condition was falsy"),
      'GenericError',
    )
  )
}
```

If this is done within a neverthrow chain which is already 2-4 intents
deep, it gets even worse.

---

I've converted the `TaggedError` type to a generic class with some
verbose & chainable helper methods

```typescript
// instead of returning `err(makeError(new Error("<message>"), "<tag>"))`
if(!condition) {
  return new TaggedError('MissingDataError')
    .withMessage('condition was falsy')
    .logged(log)
    .intoErr()
}

// `tryInto` returns a callback that checks the error instance type and
attempts to tag it
ResultAsync.fromPromise(
  RNFS.readFile(path, 'utf8'),
  new TaggedError('FileReadError').tryInto(Error)
)
Result.fromThrowable(
  () => zodSchema.parse(data),
  new TaggedError('ZodParseError').tryInto(ZodError)
)
```

Additionally, I've removed the `TagToErrorConstructorMap` map in favor
of just being able to supply a `\`${string}\`Error` tag to the
`TaggedError` constructor.