fix: prefer default room preview for announcement rooms
For a while, we've assumed that it is impossible to fetch the preview of a public broadcast room, but this has recently been fixed with the addition of the matrixPublicRoomInfo RPC, and possibly the matrix types refactor.

This PR removes some of the old hacky code we used for public broadcast rooms and implements some new slightly-less-hacky code that allows for the displaying of previews for public broadcast rooms.