fix: layout shift on PublicCommunities screen
Addresses #9499

This change removes the `useEffect` hook that was asynchronously
changing the tab on the `PublicCommunities` screen, which was causing a
race condition between the user's finger and communities loading.

With the addition of the Fedi Global Community feature on both
production and nightly, chances are that there will be at least one
public community on the Discover list.

If there ever happens to be no public communities, the user will see a
message indicating that there are no public communities.