fix: correctly check if the user's display name is not set (#9563)
* fix: correctly check if the user's display name is not set

- Addresses #9135
- Addresses #9137

Previously, we would generate a random name for the user if their npub
included their display name.

As a result, if you set your display name to something like "npub" or
"a", or any portion of the npub (e.g. "xu1dca"), then you would get a
new autogenerated name.

This PR correctly pulls the npub from the user ID before comparing it to
the user's display name.

* fix: add integration tests for matrix chat client

* chore: revert falsy check behavior for displayName