feat: immediately cancel ecash if recipient rejects foreign ecash
Addresses #8308

If you send someone foreign ecash via a chat payment and they reject the
payment, immediately cancels the ecash request and reclaims the funds.

The immediate cancellation attempt only happens on the chat screen and
does not happen automatically. If you do not check your messages and
enter the specific chat in which the recipient cancels, you will have to
wait for the original refund duration.