fix: sanitize filenames in temporary file paths before uploading (#9817)
Addresses #9621

For some reason, `react-native-fs` has an issue saving files with
commas, even when the file name is URI-encoded.

When copying a file to the temporary directory, this PR strips out all
shell-unsafe characters and commas so that it can be copied correctly.

With this change, you can now attach files with commas in the name.